Move endpoint and request for twitch to a specific python file
