+++ twitchapi/eventsub.py
@@ -7,14 +7,16 @@ from typing import Any
 from websocket import WebSocketApp
 
 from twitchapi.auth import AuthServer
-from twitchapi.utils import TwitchEndpoint, ThreadWithExc, TriggerMap, TriggerSignal, TwitchSubscriptionModel, \
+from twitchapi.twitchcom import TwitchEndpoint, TriggerSignal, TwitchSubscriptionModel, \
     TwitchSubscriptionType
+from twitchapi.utils import ThreadWithExc, TriggerMap
 from twitchapi.exception import KillThreadException, TwitchEventSubError, TwitchAuthorizationFailed
 import json
 import os
 from threading import Lock
 
-DEFAULT_DB_PATH = os.path.dirname(__file__) + "/database/TwitchDB"
+SOURCE_ROOT = os.path.dirname(__file__)
+DEFAULT_DB_PATH = SOURCE_ROOT + "/database/TwitchDB.db"
 
 
 class EventSub(WebSocketApp):
@@ -37,6 +39,9 @@ class EventSub(WebSocketApp):
 
         self.__store_in_db = store_in_db
         if self.__store_in_db:
+            if not os.path.exists(db_path):
+                self.__initialize_db(db_path=db_path)
+
             self.__db = sqlite3.connect(database=db_path, check_same_thread=False)
             self.__cursor = self.__db.cursor()
 
@@ -75,75 +80,75 @@ class EventSub(WebSocketApp):
 
                     case TwitchSubscriptionType.CHANNEL_POINT_ACTION:
                         logging.info("Process a channel point redeem")
-                        self.__process_channel_point_action(payload=payload)
+                        self.__process_channel_point_action(payload=payload, date=msg_timestamp)
 
                     case TwitchSubscriptionType.FOLLOW:
                         logging.info("Process a follow")
-                        self.__process_follow(payload=payload)
+                        self.__process_follow(payload=payload, date=msg_timestamp)
 
                     case TwitchSubscriptionType.BAN:
                         logging.info("Process a ban")
-                        self.__process_ban(payload=payload)
+                        self.__process_ban(payload=payload, date=msg_timestamp)
 
                     case TwitchSubscriptionType.SUBSCRIBE:
                         logging.info("Process a subscribe")
-                        self.__process_subscribe(payload=payload)
+                        self.__process_subscribe(payload=payload, date=msg_timestamp)
 
                     case TwitchSubscriptionType.SUBGIFT:
                         logging.info("Process a subgitf")
-                        self.__process_subgift(payload=payload)
+                        self.__process_subgift(payload=payload, date=msg_timestamp)
 
                     case TwitchSubscriptionType.RESUB_MESSAGE:
                         logging.info("Process a resub message")
-                        self.__process_resub_message(payload=payload)
+                        self.__process_resub_message(payload=payload, date=msg_timestamp)
 
                     case TwitchSubscriptionType.RAID:
                         if payload["event"]["to_broadcaster_user_id"] == self._channel_id:
                             logging.info("Process a incoming raid")
-                            self.__process_raid(payload=payload)
+                            self.__process_raid(payload=payload, date=msg_timestamp)
                         else:
                             logging.info("Process a raid")
-                            self.__process_raid_someone(payload=payload)
+                            self.__process_raid_someone(payload=payload, date=msg_timestamp)
 
                     case TwitchSubscriptionType.CHANNEL_POINT_ACTION:
                         logging.info("Process a channel action point")
-                        self.__process_channel_point_action(payload=payload)
+                        self.__process_channel_point_action(payload=payload, date=msg_timestamp)
 
                     case TwitchSubscriptionType.POLL_BEGIN:
                         logging.info("Process a poll begin")
-                        self.__process_poll_begin(payload=payload)
+                        self.__process_poll_begin(payload=payload, date=msg_timestamp)
 
                     case TwitchSubscriptionType.POLL_END:
                         logging.info("Process a poll end")
-                        self.__process_poll_end(payload=payload)
+                        self.__process_poll_end(payload=payload, date=msg_timestamp)
 
                     case TwitchSubscriptionType.PREDICTION_BEGIN:
                         logging.info("Process a prediction begin")
-                        self.__process_prediction_begin(payload=payload)
+                        self.__process_prediction_begin(payload=payload, date=msg_timestamp)
 
                     case TwitchSubscriptionType.PREDICTION_LOCK:
                         logging.info("Process a prediction lock")
-                        self.__process_prediction_lock(payload=payload)
+                        self.__process_prediction_lock(payload=payload, date=msg_timestamp)
 
                     case TwitchSubscriptionType.PREDICTION_END:
                         logging.info("Process a prediction end")
-                        self.__process_prediction_end(payload=payload)
+                        self.__process_prediction_end(payload=payload, date=msg_timestamp)
 
                     case TwitchSubscriptionType.VIP_ADD:
                         logging.info("Process a VIP added")
-                        self.__process_vip_add(payload=payload)
+                        self.__process_vip_add(payload=payload, date=msg_timestamp)
 
                     case TwitchSubscriptionType.STREAM_ONLINE:
                         logging.info("Process a stream online notification")
-                        self.__process_stream_online(payload=payload)
+                        self.__process_stream_online(payload=payload, date=msg_timestamp)
 
                     case TwitchSubscriptionType.STREAM_OFFLINE:
                         logging.info("Process a stream offline notification")
-                        self.__process_stream_offline()
+                        self.__process_stream_offline(date=msg_timestamp)
 
                     case TwitchSubscriptionType.BITS:
                         logging.info("Process bits received")
-                        self.__process_bits(payload=payload)
+                        self.__process_bits(payload=payload, date=msg_timestamp)
 
     def on_error(self, ws, message):
         logging.error(message)
@@ -152,9 +157,22 @@ class EventSub(WebSocketApp):
         logging.info("Close websocket")
         self.__thread.raise_exc(KillThreadException)
         self.__thread.join()
+        if self.__store_in_db:
+            self.__db.close()
 
     def on_open(self, ws):
         logging.info(f"Connect to {TwitchEndpoint.TWITCH_WEBSOCKET_URL}")
+
+    def __initialize_db(self, db_path:str):
+        db = sqlite3.connect(database=db_path, check_same_thread=False)
+        cursor = db.cursor()
+
+        with open(SOURCE_ROOT + "db_script/init.sql", "r") as f:
+            script = f.read()
+
+        cursor.execute(script)
+        db.commit()
+        db.close()
 
     def __subscription(self):
         for subscription in self._subscription_types:
@@ -243,7 +261,7 @@ class EventSub(WebSocketApp):
         recs = self.__lock_method(self.__cursor.execute, self.__lock_db, script)
         return recs
 
-    def __process_message(self, payload: dict[str, Any], date):
+    def __process_message(self, payload: dict[str, Any], date:str):
         event = payload["event"]
         id = event['message_id']
         user_name = event["chatter_user_login"]
@@ -259,18 +277,24 @@ class EventSub(WebSocketApp):
             self.__db__insert__message(id=id, user=user_name, message=message, date=date,
                                        parent_id=parent_id, thread_id=thread_id, cheer=cheer, emote=emote)
 
-    def __process_channel_point_action(self, payload: dict):
+    def __process_channel_point_action(self, payload: dict, date:str):
         event = payload["event"]
+        id = event["id"]
         user = event["user_name"]
         reward_name = event["reward"]["title"]
+        reward_id = event["reward"]["id"]
+        reward_title = event["reward"]["title"]
+        reward_prompt = event["reward"]["prompt"]
+        status = event["status"]
+        redeem_date = event["redeemed_at"][:-4]
         self.__trigger_map.trigger(TriggerSignal.CHANNEL_POINT_ACTION,
                                    param={"user_name": user, "reward_name": reward_name})
 
-    def __process_follow(self, payload: dict):
+    def __process_follow(self, payload: dict, date:str):
         user = payload["event"]["user_name"]
         self.__trigger_map.trigger(TriggerSignal.FOLLOW, param={"user_name": user})
 
-    def __process_subscribe(self, payload: dict):
+    def __process_subscribe(self, payload: dict, date:str):
         event = payload["event"]
         user = event["user_name"]
         tier = event["tier"]
@@ -278,7 +302,7 @@ class EventSub(WebSocketApp):
         self.__trigger_map.trigger(TriggerSignal.SUBSCRIBE, param={"user_name": user, "tier": tier,
                                                                    "is_gift": is_gift})
 
-    def __process_subgift(self, payload: dict):
+    def __process_subgift(self, payload: dict, date:str):
         event = payload["event"]
         gifter = event["user_name"]
         total = event["total"]
@@ -287,7 +311,7 @@ class EventSub(WebSocketApp):
         self.__trigger_map.trigger(TriggerSignal.SUBGIFT, param={"user_name": gifter, "tier": tier, "total": total,
                                                                  "total_gift_sub": total_gift_sub})
 
-    def __process_resub_message(self, payload: dict):
+    def __process_resub_message(self, payload: dict, date:str):
         event = payload["event"]
         user = event["user_name"]
         tier = event["tier"]
@@ -299,19 +323,19 @@ class EventSub(WebSocketApp):
                                                                        "streak": streak, "total": total,
                                                                        "duration": duration, "message": message})
 
-    def __process_raid(self, payload: dict):
+    def __process_raid(self, payload: dict, date:str):
         event = payload["event"]
         user_source = event["from_broadcaster_user_name"]
         nb_viewers = event["viewers"]
         self.__trigger_map.trigger(TriggerSignal.RAID, param={"source": user_source, "nb_viewers": nb_viewers})
 
-    def __process_raid_someone(self, payload: dict):
+    def __process_raid_someone(self, payload: dict, date:str):
         event = payload["event"]
         user_dest = event["to_broadcaster_user_name"]
         nb_viewers = event["viewers"]
         self.__trigger_map.trigger(TriggerSignal.RAID_SOMEONE, param={"dest": user_dest, "nb_viewers": nb_viewers})
 
-    def __process_poll_begin(self, payload: dict):
+    def __process_poll_begin(self, payload: dict, date:str):
         event = payload["event"]
         poll_title = event["title"]
         choices = event["choices"]
@@ -324,7 +348,7 @@ class EventSub(WebSocketApp):
                                                                     "channel_point_settings": channel_point_settings,
                                                                     "start_date": start, "end_date": end})
 
-    def __process_poll_end(self, payload: dict):
+    def __process_poll_end(self, payload: dict, date:str):
         event = payload["event"]
         poll_title = event["title"]
         choices = event["choices"]
@@ -332,7 +356,7 @@ class EventSub(WebSocketApp):
         self.__trigger_map.trigger(TriggerSignal.POLL_END, param={"title": poll_title, "choices": choices,
                                                                   "status": status})
 
-    def __process_prediction_begin(self, payload: dict):
+    def __process_prediction_begin(self, payload: dict, date:str):
         event = payload["event"]
         pred_title = event["title"]
         choices = event["outcomes"]
@@ -341,13 +365,13 @@ class EventSub(WebSocketApp):
         self.__trigger_map.trigger(TriggerSignal.PREDICTION_BEGIN, param={"title": pred_title, "choices": choices,
                                                                           "start_date": start, "lock_date": lock})
 
-    def __process_prediction_lock(self, payload: dict):
+    def __process_prediction_lock(self, payload: dict, date:str):
         event = payload["event"]
         pred_title = event["title"]
         result = event["outcomes"]
         self.__trigger_map.trigger(TriggerSignal.PREDICTION_BEGIN, param={"title": pred_title, "result": result})
 
-    def __process_prediction_end(self, payload: dict):
+    def __process_prediction_end(self, payload: dict, date:str):
         event = payload["event"]
         pred_title = event["title"]
         result = event["outcomes"]
@@ -361,7 +385,7 @@ class EventSub(WebSocketApp):
         self.__trigger_map.trigger(TriggerSignal.PREDICTION_BEGIN, param={"title": pred_title, "result": result,
                                                                           "winning_pred": winning})
 
-    def __process_ban(self, payload: dict):
+    def __process_ban(self, payload: dict, date:str):
         event = payload["event"]
         user = event["user_name"]
         reason = event["reason"]
@@ -371,21 +395,21 @@ class EventSub(WebSocketApp):
         self.__trigger_map.trigger(TriggerSignal.BAN, param={"user_name": user, "reason": reason, "start_ban": ban_date,
                                                              "end_ban": end_ban, "permanent": permanent})
 
-    def __process_vip_add(self, payload: dict):
+    def __process_vip_add(self, payload: dict, date:str):
         event = payload["event"]
         user = event["user_name"]
         self.__trigger_map.trigger(TriggerSignal.VIP_ADD, param={"user_name": user})
 
-    def __process_stream_online(self, payload: dict):
+    def __process_stream_online(self, payload: dict, date:str):
         event = payload["event"]
         type = event["type"]
         start = event["started_at"]
         self.__trigger_map.trigger(TriggerSignal.STREAM_ONLINE, param={"type": type, "start_time": start})
 
-    def __process_stream_offline(self):
+    def __process_stream_offline(self, date:str):
         self.__trigger_map.trigger(TriggerSignal.STREAM_OFFLINE)
 
-    def __process_bits(self, payload: dict):
+    def __process_bits(self, payload: dict, date:str):
         event = payload["event"]
         user = event["user_name"]
         bits_number = event["bits"]
